fix(contracts): prevent return data gas grief (#1099)
Prevent return data mem copy gas grief.

- remove custom "return data too long" error

task: https://app.asana.com/0/1206208509925075/1207375377017425

+++ contracts/src/interfaces/IOmniPortal.sol
@@ -72,7 +72,7 @@ interface IOmniPortal {
     /**
      * @notice Maxium number of bytes allowed in xreceipt result
      */
-    function xreceiptMaxErrorBytes() external view returns (uint64);
+    function xreceiptMaxErrorBytes() external view returns (uint16);
 
     /**
      * @notice Returns the chain ID of the chain to which this portal is deployed
